fix all -Wall -Wextra warnings in forward/ and tests/
forward/src/element.cpp: remove unused n_node, l2mu
forward/src/pml.cpp:   comment out unused u param name
forward/src/io.cpp:    comment out unused rank param name
forward/src/solver.cpp: comment out unused beta param name in newmark_correct
compress/include/gf/PrecisionPolicy.h: constexpr→inline for HDF5 type helpers
tests/test_io.cpp:     static_cast for signedness comparison
tests/test_source.cpp: remove unused n_dof
tests/test_gll.cpp:    remove unused ngl

+++ tests/test_io.cpp
@@ -109,7 +109,7 @@ TEST_CASE("Read partition data round-trip", "[io]") {
     REQUIRE(data.n_local_elem == n_local);
     REQUIRE(data.n_ghost_elem == 0);
     REQUIRE(data.ngll == ngll);
-    REQUIRE(data.local_element_ids.size() == n_local);
+    REQUIRE(data.local_element_ids.size() == static_cast<size_t>(n_local));
     REQUIRE(data.ghost_element_ids.empty());
     REQUIRE(data.ghost_owners.empty());
